refact: refactored code, removed unused imports, renamed methods and packages to match the patterns specified by the team
