+++ MyMusic/src/main/java/com/ciandt/summit/bootcamp2022/application/adapters/controllers/SongsController.java
@@ -9,6 +9,7 @@ import com.ciandt.summit.bootcamp2022.domains.songs.ports.interfaces.SongService
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -16,7 +17,7 @@ import org.springframework.web.bind.annotation.*;
 @RequestMapping("/api")
 public class SongsController implements SongControllerDocs {
 
-    private static Logger logger = LoggerFactory.getLogger(SongsController.class.getSimpleName());
+    private static final Logger logger = LoggerFactory.getLogger(SongsController.class.getSimpleName());
 
     @Autowired
     private SongServicePort songServicePort;
@@ -26,12 +27,17 @@ public class SongsController implements SongControllerDocs {
     public ResponseEntity<SongResponseDTO> findSongsByNameOrArtistName(@RequestParam(name = "filtro", required = false) String filter,
                                                                        @RequestParam(name = "pagina", defaultValue = "0") int pageNumber)
             throws SongsNotFoundException, InvalidSongNameOrArtistNameException {
-            logger.info("Recebendo Request Get para /musicas");
+
+        logger.info("Recebendo Request Get para /musicas");
         SongResponseDTO response;
         if(filter == null){
             response = songServicePort.findAllSongs(pageNumber);
         } else {
-             response = songServicePort.findByNameOrArtistName(filter, pageNumber);
+            response = songServicePort.findByNameOrArtistName(filter, pageNumber);
+        }
+
+        if(response.getData().isEmpty()){
+            return new ResponseEntity<>(response, HttpStatus.NO_CONTENT);
         }
 
         return ResponseEntity.ok().body(response);
