--- MyMusic/src/main/java/com/ciandt/summit/bootcamp2022/domains/exceptions/songs/playlists/PlaylistsNotFoundException.java
@@ -1,4 +1,0 @@
-package com.ciandt.summit.bootcamp2022.domains.exceptions.songs.playlists;
-
-public class PlaylistsNotFoundException {
-}
