+++ MyMusic/src/main/java/com/ciandt/summit/bootcamp2022/domains/exceptions/tokens/BadAuthRequestException.java
@@ -4,8 +4,4 @@ public class BadAuthRequestException extends RuntimeException {
     public BadAuthRequestException(String message) {
         super(message);
     }
-
-    public BadAuthRequestException(String message, Throwable cause) {
-        super(message, cause);
-    }
 }
