+++ MyMusic/src/main/java/com/ciandt/summit/bootcamp2022/domains/exceptions/tokens/UnauthorizedException.java
@@ -4,8 +4,4 @@ public class UnauthorizedException extends RuntimeException {
     public UnauthorizedException(String message) {
         super(message);
     }
-
-    public UnauthorizedException(String message, Throwable cause) {
-        super(message, cause);
-    }
 }
