+++ MyMusic/src/main/java/com/ciandt/summit/bootcamp2022/domains/songs/ports/interfaces/SongServicePort.java
@@ -2,10 +2,7 @@ package com.ciandt.summit.bootcamp2022.domains.songs.ports.interfaces;
 
 import com.ciandt.summit.bootcamp2022.domains.exceptions.songs.InvalidSongNameOrArtistNameException;
 import com.ciandt.summit.bootcamp2022.domains.exceptions.songs.SongsNotFoundException;
-import com.ciandt.summit.bootcamp2022.domains.songs.dtos.SongDTO;
 import com.ciandt.summit.bootcamp2022.domains.songs.dtos.SongResponseDTO;
-
-import java.util.List;
 
 public interface SongServicePort {
     SongResponseDTO findByNameOrArtistName(String name, int pageNumber)
